// marketDataService/src/enums/message.ts
export enum ErrorMessages {
  INTERNAL_SERVER_ERROR = "Internal Server Error",
  TOKEN_NOT_PROVIDED = "Authentication token not provided",
  INVALID_TOKEN = "Invalid or malformed token",
  TOKEN_EXPIRED = "Authentication token expired",
  FORBIDDEN = "Forbidden: Insufficient privileges",
  GENERIC_ERROR = "An unknown error occurred.",
  INVALID_SERVICE_ID = "Unidentified Service",
}